requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-filter==23.5
Pillow==10.2.0
mysqlclient==2.2.4
python-dotenv==1.0.1
stripe==7.11.0
boto3==1.34.34
django-storages==1.14.2